word delete dialog, snacbar and update word dialog components added
any word can be updated and deleted on redux

+++ src/components/WordCard/index.js
@@ -2,12 +2,15 @@ import React, { useEffect, useState } from 'react'
 import Box3d from './Box3d'
 import { setAllSwipeRotation, setOneSwipeRotation, deleteWord } from "../../app/Slices/WordsSlice"
 import { useDispatch } from 'react-redux'
-export default function WordCard({ word, translation, swipeRotation, id }) {
+import AlertDialog from '../common/AlertDialog'
+import UpdateDialog from './UpdateDialog'
+export default function WordCard({ word, translation, swipeRotation, id, snackBarOpen }) {
     const [rotate, setRotate] = useState({ x: 0, y: 0, z: 0 })
     const [swipeStart, setSwipeStart] = useState(null)
     const [swipeEnd, setSwipeEnd] = useState(null)
     const [faceName, setFaceName] = useState("front")//front, back, top, bottom
-
+    const [isAlertOpen, setIsAlertOpen] = useState(false);
+    const [isUpdateOpen, setIsUpdateOpen] = useState(false);
     const dispatch = useDispatch()
 
     const setSwipeRotation = (rotation, id) => {
@@ -43,7 +46,6 @@ export default function WordCard({ word, translation, swipeRotation, id }) {
     }, [swipeStart, swipeEnd])
 
 
-
     useEffect(() => {
 
         console.log("use effect");
@@ -54,17 +56,17 @@ export default function WordCard({ word, translation, swipeRotation, id }) {
             console.log("r left");
             if (faceName == "front") setRotate({ x: rotate.x, y: rotate.y - 45, z: rotate.z })
             if (faceName == "bottom") setRotate({ x: rotate.x, y: rotate.y, z: rotate.z + 45 })
-            if (faceName == "back") setRotate({ x: rotate.x, y: rotate.y+45, z: rotate.z })
+            if (faceName == "back") setRotate({ x: rotate.x, y: rotate.y + 45, z: rotate.z })
             if (faceName == "top") setRotate({ x: rotate.x, y: rotate.y, z: rotate.z - 45 })
         } else if (swipeRotation === "right") {
             console.log("r right");
             if (faceName == "front") setRotate({ x: rotate.x, y: rotate.y + 45, z: rotate.z })
-            if (faceName == "bottom") setRotate({ x: rotate.x, y: rotate.y, z: rotate.z  - 45 })
-            if (faceName == "back") setRotate({ x: rotate.x, y: rotate.y-45, z: rotate.z })
+            if (faceName == "bottom") setRotate({ x: rotate.x, y: rotate.y, z: rotate.z - 45 })
+            if (faceName == "back") setRotate({ x: rotate.x, y: rotate.y - 45, z: rotate.z })
             if (faceName == "top") setRotate({ x: rotate.x, y: rotate.y, z: rotate.z + 45 })
         } else if (swipeRotation === "center") {
             console.log("r center");
-            setRotate({ x: rotate.x, y: 0, z: 0})
+            setRotate({ x: rotate.x, y: 0, z: 0 })
         }
     }, [swipeRotation])
     return (
@@ -76,17 +78,28 @@ export default function WordCard({ word, translation, swipeRotation, id }) {
 
 
         }} style={{ width: '100%' }}>
-
+            <UpdateDialog  open={isUpdateOpen} confirm={()=>{
+                setIsUpdateOpen(false);
+                snackBarOpen();
+            }} data={{id,word,translation}} cancel={()=>{ setIsUpdateOpen(false) }}  />
+            <AlertDialog title="Are you sure ?" content={`Do you want to delete this word? ${word} = ${translation}`} confirm={() => { 
+                dispatch(deleteWord({ id })); 
+                setIsAlertOpen(false); 
+                snackBarOpen();
+                }} cancel={() => { setIsAlertOpen(false) }} open={isAlertOpen} />
             <Box3d onTouchEnd={(e) => {
                 //console.log("End ",e)
                 setSwipeEnd(e.changedTouches[0].clientX)
             }} onTouchStart={(e) => {
                 // console.log("Start ",e);
                 setSwipeStart(e.changedTouches[0].clientX)
-            }} onDelete={(e)=>{
-                e.stopPropagation();   
-                dispatch(deleteWord({id}))       
-            }}  faces={{ front: word, bottom: translation, back: word, top: translation }} rotate={rotate} />
+            }} onDelete={(e) => {
+                e.stopPropagation();
+                setIsAlertOpen(true);
+            }} onUpdate={(e)=>{
+                e.stopPropagation();
+                setIsUpdateOpen(true);
+            }} faces={{ front: word, bottom: translation, back: word, top: translation }} rotate={rotate} />
         </div>
     )
 }
